budget-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-dotenv
psycopg2
python-multipart